+++ src/main/java/vendingmachine/Application.java
@@ -5,7 +5,8 @@ import vendingmachine.controller.VendingMachineInputController;
 public class Application {
 	public static void main(String[] args) {
 		// TODO: 프로그램 구현
-		VendingMachineInputController.setCoin();
+		int vendingMachineAccount = VendingMachineInputController.getAccount();
+		System.out.println("vendingMachineAccount = " + vendingMachineAccount);
 
 	}
 }
