+++ src/main/java/vendingmachine/service/InputExceptionService.java
@@ -1,6 +1,6 @@
-package vendingmachine.controller;
+package vendingmachine.service;
 
-public class ExceptionController {
+public class InputExceptionService {
 	private static final String NOT_MONEY_ERROR_MESSAGE = "10으로 나누어 떨어지는 0 이상의 정수여야 합니다.";
 	private static final String NOT_INTEGER_ERROR_MESSAGE = "숫자가 아닙니다.";
